Authentication API implementation with Test results

+++ src/main/java/com/innovatehub/order/proxy/OrderServiceProxyImpl.java
@@ -1,5 +1,7 @@
 package com.innovatehub.order.proxy;
 
+import com.innovatehub.cache.APIAuthorizeCache.APICacheManager;
+import com.innovatehub.exception.APISecurityException;
 import com.innovatehub.order.dao.Order;
 import com.innovatehub.order.dto.OrderResponse;
 import com.innovatehub.order.service.OrderService;
@@ -8,7 +10,7 @@ import org.springframework.http.HttpStatus;
 
 import java.util.List;
 
-public class OrderServiceProxyImpl implements OrderService {
+public class OrderServiceProxyImpl implements OrderProxyService {
 
     private OrderService orderService;
     private OrderValidator orderValidator;
@@ -19,22 +21,42 @@ public class OrderServiceProxyImpl implements OrderService {
     }
 
     @Override
-    public void updateOrder(Order order) {
+    public void updateOrder(Order order, String tokenID)throws APISecurityException {
+
+        if(!APICacheManager.findRoleHasAPIAcesss(tokenID, "order")) {
+            throw new APISecurityException("Unauthorized operation");
+        }
+
         orderService.updateOrder(order);
     }
 
     @Override
-    public List<Order> findAllOrder() {
+    public List<Order> findAllOrder(String tokenID)throws APISecurityException {
+
+        if(!APICacheManager.findRoleHasAPIAcesss(tokenID, "order")) {
+            throw new APISecurityException("Unauthorized operation");
+        }
+
         return orderService.findAllOrder();
     }
 
     @Override
-    public void deleteByOrderId(int OrderID) {
+    public void deleteByOrderId(int OrderID, String tokenID)throws APISecurityException {
+
+        if(!APICacheManager.findRoleHasAPIAcesss(tokenID, "order")) {
+            throw new APISecurityException("Unauthorized operation");
+        }
+
         orderService.deleteByOrderId(OrderID);
     }
 
     @Override
-    public Order saveOrder(Order order) {
+    public Order saveOrder(Order order, String tokenID)throws APISecurityException {
+
+        if(!APICacheManager.findRoleHasAPIAcesss(tokenID, "order")) {
+            throw new APISecurityException("Unauthorized operation");
+        }
+
         if (!orderValidator.isValid(order)) {
             return new OrderResponse("Invalid  order request", order, HttpStatus.BAD_REQUEST);
         }
