+++ src/main/java/com/innovatehub/user/proxy/UserServiceProxyImpl.java
@@ -1,5 +1,7 @@
 package com.innovatehub.user.proxy;
 
+import com.innovatehub.cache.APIAuthorizeCache.APICacheManager;
+import com.innovatehub.exception.APISecurityException;
 import com.innovatehub.user.dao.Member;
 import com.innovatehub.user.dto.UserResponse;
 import com.innovatehub.user.service.UserService;
@@ -8,7 +10,7 @@ import org.springframework.http.HttpStatus;
 
 import java.util.List;
 
-public class UserServiceProxyImpl implements UserService {
+public class UserServiceProxyImpl implements UserServiceProxy {
 
     private UserService userService;
     private UserValidator userValidator;
@@ -19,22 +21,42 @@ public class UserServiceProxyImpl implements UserService {
     }
 
     @Override
-    public void updateMember(Member member) {
+    public void updateMember(Member member, String tokenID)throws APISecurityException {
+
+        if(!APICacheManager.findRoleHasAPIAcesss(tokenID, "user")) {
+            throw new APISecurityException("Unauthorized operation");
+        }
+
         userService.updateMember(member);
     }
 
     @Override
-    public List<Member> findAll() {
+    public List<Member> findAll(String tokenID)throws APISecurityException {
+
+        if(!APICacheManager.findRoleHasAPIAcesss(tokenID, "user")) {
+            throw new APISecurityException("Unauthorized operation");
+        }
+
         return userService.findAll();
     }
 
     @Override
-    public void deleteByMemberId(int MemberID) {
+    public void deleteByMemberId(int MemberID, String tokenID)throws APISecurityException {
+
+        if(!APICacheManager.findRoleHasAPIAcesss(tokenID, "user")) {
+            throw new APISecurityException("Unauthorized operation");
+        }
+
         userService.deleteByMemberId(MemberID);
     }
 
     @Override
-    public Member saveMember(Member member) {
+    public Member saveMember(Member member, String tokenID)throws APISecurityException {
+
+        if(!APICacheManager.findRoleHasAPIAcesss(tokenID, "user")) {
+            throw new APISecurityException("Unauthorized operation");
+        }
+
         if (!userValidator.isValid(member)) {
             return new UserResponse(member, "User is not valid", HttpStatus.BAD_REQUEST);
         }
